Split monolith into reasonable modules in preparation for further refactoring
