+++ one-dimensional/ui.py
@@ -2,73 +2,17 @@ import json
 import os
 import uuid
 import random
+import statistics
+import time
+
 import matplotlib.pyplot as plt
 from matplotlib.ticker import MaxNLocator
-import time
 import tkinter as tk
 from tkinter import ttk, messagebox, filedialog
-import statistics
+
 from placement_optimization_sim import NumberLine
-
-
-class Simulation:
-    def __init__(self, number_line, iterations=1, repetitions=1, significant_figures=1, progress_callback=None):
-        self.number_line = number_line
-        self.iterations = iterations
-        self.repetitions = repetitions
-        self.significant_figures = significant_figures
-        self.optimal_p_values = []
-        self.progress_callback = progress_callback
-
-    def run(self):
-        for _ in range(self.repetitions):
-            p_val = self.__funnel_to_p_value()
-            self.optimal_p_values.append(p_val)
-            if self.progress_callback:
-                self.progress_callback()
-
-    def noprogress(self):
-        pass
-
-    def __gather(self):
-        dataset = []
-        for _ in range(self.iterations):
-            dataset.append(self.number_line.regenerate_data())
-        return statistics.mean(dataset)
-
-    # ! Definite Bottleneck
-    def __funnel_to_p_value(self):
-        left_bound = float(self.number_line.get_starting_position())
-        right_bound = float(self.number_line.get_end())
-        step = 1
-        # ! Space for data structure improvement here, lists are not the best
-        traversal_distances = []
-        tested_p_values = []
-
-        for _ in range(self.significant_figures):
-            traversal_distances.clear()
-            tested_p_values.clear()
-            step /= 10
-            j = left_bound
-            while j <= right_bound:
-                self.number_line.set_starting_position(j)
-                traversal = self.__gather()
-                traversal_distances.append(traversal)
-                tested_p_values.append(j)
-                j += float(step)
-
-            optimal_p_val = self.__find_optimal_p(
-                traversal_distances, tested_p_values)
-            end = self.number_line.get_end()
-            left_bound = min(optimal_p_val - step, end)
-            right_bound = min(optimal_p_val + step, end)
-
-        return self.__find_optimal_p(traversal_distances, tested_p_values)
-
-    def __find_optimal_p(self, traversal_distances, tested_p_values):
-        minimum_traversal = min(traversal_distances)
-        idx = traversal_distances.index(minimum_traversal)
-        return tested_p_values[idx]
+from utils import ProgramTimer, ProgressBar
+from simulation import Simulation
 
 
 class UserInterface:
@@ -138,13 +82,14 @@ class UserInterface:
             "Round final stats stdev to places: ", self.stdev_decimal_places, 1, 0, master=self.stats_frame, width=2)
 
         # Buttons and progress
+
         self.run_button = ttk.Button(
             self.root, text="Run", command=self.__try_run_simulation_with_single_plot)
-        self.run_button.grid(row=5, column=1, padx=10, pady=10)
+        self.run_button.grid(row=5, column=0, padx=10, pady=10)
 
-        self.quit_button = ttk.Button(
-            self.root, text="Quit", command=self.quit_app)
-        self.quit_button.grid(row=5, column=4, padx=10, pady=10)
+        self.replot_button = ttk.Button(
+            self.root, text="Replot", command=self.try_replot)
+        self.replot_button.grid(row=5, column=1, padx=10, pady=10)
 
         self.export_button = ttk.Button(
             self.root, text="Export Run", command=self.try_export)
@@ -155,6 +100,10 @@ class UserInterface:
         self.import_button.grid(row=5, column=3, padx=10, pady=10)
 
         self.progress_bar = ProgressBar(self.root, bar_row=6, label_row=6)
+
+        self.quit_button = ttk.Button(
+            self.root, text="Quit", command=self.quit_app)
+        self.quit_button.grid(row=5, column=4, padx=10, pady=10)
 
         self.recalculate_stats_button = ttk.Button(
             self.stats_frame, text="Recalculate", command=self.__calculate_stats_for_superset)
@@ -187,6 +136,14 @@ class UserInterface:
         )
         if file_path:
             self.__import_data(file_path)
+
+    def try_replot(self):
+        if len(self.optimal_distance_from_center_superset) > 0:
+            self.__plot_optimal_distances()
+        else:
+            # ! Replace with elegent error handling
+            messagebox.showwarning(
+                title="Replot Failed", message="Cannot replot data that has not been generated. Run a simulation first.")
 
     def __create_label_and_entry(self, text, variable, row, col, scale_to=None, master=None, width=None, focus=False):
         m = master or self.root
@@ -311,7 +268,7 @@ class UserInterface:
         with open(path, 'x') as f:
             f.write(json_data)
             messagebox.showinfo(title="Export completed",
-                                message=f"Exported as:\n{filename}.")
+                                message=f"Exportd as:\n{filename}.")
 
     def __import_data(self, path: str):
         with open(path, 'r') as f:
@@ -368,88 +325,3 @@ class UserInterface:
 
     def quit_app(self):
         self.root.quit()
-
-
-class ProgramTimer:
-    def __init__(self):
-        self.init_time = time.time()
-        self.counter = time.time()
-        self.start_time = None
-
-    def start(self):
-        self.start_time = time.time()
-
-    def get_time_since_start(self):
-        return time.time() - self.start_time
-
-    def get_time_since_init(self):
-        return time.time() - self.init_time
-
-    def get_counter_time(self):
-        return time.time() - self.counter
-
-    def reset_counter(self, step=None):
-        print(f"Counter reset on {step}")
-        self.counter = time.time()
-
-    def report_step(self, step):
-        now_time = time.time()
-        i_time = now_time - self.init_time
-        s_time = now_time - (self.start_time or now_time)
-        c_time = now_time - self.counter
-        print(f"Step {step} report:")
-        print(f"Time since init: {i_time:.2f}")
-        print(f"Time since start: {s_time:.2f}")
-        print(f"Time since counter reset: {c_time:.2f}")
-
-
-class ProgressBar:
-    def __init__(self, master, max_count=100, current_count=0, bar_row=0, label_row=0, bar_col=0, label_col=3):
-        self.max_count = max_count
-        self.current_count = current_count
-        self.master = master
-
-        self.progress_bar = ttk.Progressbar(
-            master, orient='horizontal', mode='determinate')
-        self.progress_bar.grid(row=bar_row, column=bar_col, columnspan=3,
-                               padx=10, pady=10, sticky='ew')
-
-        percentage = self.__get_current_percent()
-        self.progress_label = tk.Label(
-            master, text=f'{percentage}%')
-        self.progress_label.grid(
-            row=label_row, column=label_col, padx=10, pady=10, sticky='w')
-
-        self.__update()
-
-    def __get_current_percent(self):
-        return int((self.current_count / self.max_count) * 100)
-
-    def update_progress(self, count=None, max_count=None):
-        if max_count:
-            self.max_count = max_count
-        if count:
-            self.current_count = count
-        self.__update()
-
-    def increment_progress(self):
-        self.current_count += 1
-        self.__update()
-
-    def clear_progress(self):
-        self.current_count = 0
-        self.__update()
-
-    def __update(self):
-        percent = self.__get_current_percent()
-        self.progress_bar['maximum'] = self.max_count
-        self.progress_bar['value'] = self.current_count
-        self.progress_label.config(text=f'{percent}%')
-        self.master.update_idletasks()
-
-
-if __name__ == "__main__":
-    root = tk.Tk()
-    timer = ProgramTimer()
-    app = UserInterface(root, timer)
-    root.mainloop()
